Move SearchBar into separate page
